+++ cq/storages.py
@@ -27,7 +27,7 @@ class Storage:
             revision=revision,
         )
         event = self.append(event)
-        handle_event(event)
+        handle_event(event, replaying_events=False)
         return event
 
     def create_event(self, id, aggregate_type, name, aggregate_id, data=None, ts=None, revision=1):
@@ -58,6 +58,10 @@ class Storage:
 
     def has_unique(self, namespace, value):
         raise NotImplementedError
+
+    def replay_events(self):
+        for event in self.iter_all_events():
+            handle_event(event, replaying_events=True)
 
 
 class LocalMemoryStorage(Storage):