+++ examples/djangoapp/accounts/handlers.py
@@ -1,10 +1,9 @@
-from .aggregates import User
 from .models import User as UserProjection
 from cq.handlers import register_handler
 
 
 @register_handler('User', 'Registered')
-def handle_registered_user(event):
+def handle_registered_user(event, replaying_events):
     UserProjection.objects.create(
         id=event.aggregate_id,
         email=event.data['email'],